requirements.txt
APScheduler==3.6.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.0.6
Flask==1.0.2
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
gunicorn==19.9.0
idna==2.8
ipython==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pexpect==4.0.1
pkg-resources==0.0.0
ptyprocess==0.5
pyee==5.0.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
simplegeneric==0.8.1
six==1.12.0
slackclient==1.3.1
slackeventsapi==2.1.0
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.4.3
websocket-client==0.54.0
Werkzeug==0.15.2
zdesk==2.7.1
